Modified Kappa-fitting EISPAC (0.95.0) directory folder
Contains the eispac version 0.95.0 directory, modified to allow the use of kappa fitting functions for the spectra. The directory itself is still named 'eispac' as in the default installation.
This is somewhat of a placeholder, planning to change the directory with its own name and modifying the files to allow it to be imported under a separate name to distinguish it from the default installation.