Main plotting and fitting script directory
This contains a number of directories required to save plots and data arrays if required. This also contains the main plotting and fitting script (FittingMultiFile.py) which can automatically process multiple data (+header) files as provided within the 'filequeue' folder.

Also included is a script for comparing the summary dataset files with the GOES X-ray curve. The data for this is provided in the Kappa1 and Kappa 2 directories (for 2016 and 2017 Jeffrey N. Kappa function respectively)